Put version.json into CARGO_MANIFEST_DIR regardless of cargo profile.
I was having trouble where the file wasn't there, maybe because I had built
with the release profile first and then with the debug profile and build.rs
wasn't rerun because it didn't explicitly depend on the PROFILE env var.

Let's just always place the file in the same spot regardless of cargo profile.
The file is on the .gitignore so it shouldn't be a problem to have it in the root dir.